backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "akumaluffy", email = "yliu0602@gmail.com"},
]
dependencies = ["fastapi>=0.121.0", "uvicorn>=0.38.0", "openai>=1.0.0", "python-dotenv>=1.0.0"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "uvicorn app.main:app --reload --port 8000"